Refactor click actions, allow any language for scripting
