+++ .gitignore
@@ -1,2 +1,5 @@
 *.o
 main
+click-actions/raise
+click-actions/minimize
+click-actions/close
