--- click-actions/close
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-"$(dirname "$0")"/../main --close "$1"
