--- click-actions/minimize
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-"$(dirname "$0")"/../main --minimize "$1"
