--- click-actions/raise
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-"$(dirname "$0")"/../main --raise "$1"
