+++ windowlist.c
@@ -22,39 +22,6 @@
 
 #define MAX_PROPERTY_VALUE_LEN 4096
 
-int client_msg(Display* d, Window w, char* msg) {
-    XEvent e;
-    long mask = SubstructureRedirectMask | SubstructureNotifyMask;
-
-    e.xclient.type = ClientMessage;
-    e.xclient.serial = 0;
-    e.xclient.send_event = True;
-    e.xclient.message_type = XInternAtom(d, msg, False);
-    e.xclient.window = w;
-    e.xclient.format = 32;
-
-    if (XSendEvent(d, DefaultRootWindow(d), False, mask, &e)) {
-        return EXIT_SUCCESS;
-    } else {
-        fprintf(stderr, "Cannot send %s event.\n", msg);
-        return EXIT_FAILURE;
-    }
-}
-
-int close_window(Display* d, Window w) {
-    return client_msg(d, w, "_NET_CLOSE_WINDOW");
-}
-
-int raise_window(Display* d, Window w) {
-    client_msg(d, w, "_NET_ACTIVE_WINDOW");
-    XMapRaised(d, w);
-    return EXIT_SUCCESS;
-}
-
-int minimize_window(Display* d, Window w) {
-    return !XIconifyWindow(d, w, DefaultScreen(d));
-}
-
 char* get_property(Display* d, Window w, Atom xa_prop_type, char* prop_name,
                    unsigned long* size) {
     unsigned long ret_nitems, ret_bytes_after, tmp_size;
