+++ windowlist.h
@@ -10,6 +10,3 @@ struct window_props {
 struct window_props* generate_window_list(Display* d, long current_desktop_id, int* n_wprops);
 Window get_active_window(Display* d);
 long get_desktop_id(Display* d, Window w, char* prop_name);
-int close_window(Display* d, Window w);
-int raise_window(Display* d, Window w);
-int minimize_window(Display* d, Window w);
